requirements.txt
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
certifi==2021.10.8
decorator==5.1.0
Django==3.2.9
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.4
drf-spectacular==0.21.0
gunicorn==20.1.0
importlib-resources==5.4.0
inflection==0.5.1
ipython==7.29.0
jedi==0.18.1
jsonschema==4.2.1
matplotlib-inline==0.1.3
mpmath==1.2.1
mysqlclient==2.0.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
psycopg2==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
sympy==1.9
traitlets==5.1.1
uritemplate==4.1.1
wcwidth==0.2.5
zipp==3.6.0
